Extract shared autocomplete unit-test helpers into helpers.js
